# make ivy-bibtex faster by only formatting displayed candidates