--- a/bibtex-completion.el
+++ b/bibtex-completion.el
@@ -522,53 +522,32 @@ values."
                concat (car p))
     nil))
 
-(defun bibtex-completion-normalize-candidate (candidate)
-  "Extract key from CANDIDATE."
-  (cond ((stringp candidate)
-         candidate)
-        ((stringp (cdr-safe candidate))
-         (cdr candidate))))
-
-(defun bibtex-completion-normalize-candidates (candidates)
-  "Normalize CANDIDATES to a list of candidates for each action
-function."
-  (cond
-   ;; single cons cell
-   ((stringp (cdr-safe candidates))
-    (list (cdr candidates)))
-   ((listp candidates)
-    (-map #'bibtex-completion-normalize-candidate candidates))
-   (t
-    (list candidates))))
-
 
-(defun bibtex-completion-open-pdf (candidates)
+(defun bibtex-completion-open-pdf (keys)
   "Open the PDFs associated with the marked entries using the
 function specified in `bibtex-completion-pdf-open-function'.  All paths
 in `bibtex-completion-library-path' are searched.  If there are several
 matching PDFs for an entry, the first is opened."
   (--if-let
       (-flatten
-       (-map 'bibtex-completion-find-pdf
-             (bibtex-completion-normalize-candidates candidates)))
+       (-map 'bibtex-completion-find-pdf keys))
       (-each it bibtex-completion-pdf-open-function)
     (message "No PDF(s) found.")))
 
-(defun bibtex-completion-open-url-or-doi (candidates)
+(defun bibtex-completion-open-url-or-doi (keys)
   "Open the associated URL or DOI in a browser."
-  (let ((keys (bibtex-completion-normalize-candidates candidates)))
-    (dolist (key keys)
-      (let* ((entry (bibtex-completion-get-entry key))
-             (url (bibtex-completion-get-value "url" entry))
-             (doi (bibtex-completion-get-value "doi" entry))
-             (browse-url-browser-function
-              (or bibtex-completion-browser-function
-                  browse-url-browser-function)))
-        (if url (browse-url url)
-          (if doi (browse-url
-                   (s-concat "http://dx.doi.org/" doi)))
-          (message "No URL or DOI found for this entry: %s"
-                   key))))))
+  (dolist (key keys)
+    (let* ((entry (bibtex-completion-get-entry key))
+           (url (bibtex-completion-get-value "url" entry))
+           (doi (bibtex-completion-get-value "doi" entry))
+           (browse-url-browser-function
+            (or bibtex-completion-browser-function
+                browse-url-browser-function)))
+      (if url (browse-url url)
+        (if doi (browse-url
+                 (s-concat "http://dx.doi.org/" doi)))
+        (message "No URL or DOI found for this entry: %s"
+                 key)))))
 
 (defun bibtex-completion-format-citation-default (keys)
   "Default formatter for keys, separates multiple keys with commas."
@@ -657,20 +636,18 @@ omitted."
                 for pdfs = (bibtex-completion-find-pdf key)
                 append (--map (format "[[%s][%s]]" it key) pdfs))))
 
-(defun bibtex-completion-insert-citation (candidates)
+(defun bibtex-completion-insert-citation (keys)
   "Insert citation at point.  The format depends on
 `bibtex-completion-format-citation-functions'."
-  (let ((keys (bibtex-completion-normalize-candidates candidates))
-        (format-function
+  (let ((format-function
          (cdr (or (assoc major-mode bibtex-completion-format-citation-functions)
                   (assoc 'default   bibtex-completion-format-citation-functions)))))
     (insert
      (funcall format-function keys))))
 
-(defun bibtex-completion-insert-reference (candidates)
+(defun bibtex-completion-insert-reference (keys)
   "Insert a reference for each selected entry."
-  (let* ((keys (bibtex-completion-normalize-candidates candidates))
-         (refs (--map
+  (let* ((refs (--map
                 (s-word-wrap fill-column
                              (concat "\n- " (bibtex-completion-apa-format-reference it)))
                 keys)))
@@ -680,9 +657,7 @@ omitted."
   "Returns a plain text reference in APA format for the
 publication specified by KEY."
   (let*
-   ((key (bibtex-completion-normalize-candidate key))
-    (entry (bibtex-completion-get-entry
-            (bibtex-completion-normalize-candidate key)))
+   ((entry (bibtex-completion-get-entry key))
     (ref (pcase (downcase (bibtex-completion-get-value "=type=" entry))
            ("article"
             (s-format
@@ -826,20 +801,17 @@ defined.  Surrounding curly braces are stripped."
           value))
       default)))
 
-(defun bibtex-completion-insert-key (candidates)
+(defun bibtex-completion-insert-key (keys)
   "Insert BibTeX key at point."
-  (let ((keys (bibtex-completion-normalize-candidates candidates)))
-    (insert
-     (funcall 'bibtex-completion-format-citation-default keys))))
+  (insert
+   (funcall 'bibtex-completion-format-citation-default keys)))
 
-(defun bibtex-completion-insert-bibtex (candidates)
+(defun bibtex-completion-insert-bibtex (keys)
   "Insert BibTeX key at point."
-  (let ((keys (bibtex-completion-normalize-candidates candidates)))
-    (insert (s-join "\n" (--map (bibtex-completion-make-bibtex it) keys)))))
+  (insert (s-join "\n" (--map (bibtex-completion-make-bibtex it) keys))))
 
 (defun bibtex-completion-make-bibtex (key)
-  (let* ((entry (bibtex-completion-get-entry
-                 (bibtex-completion-normalize-candidate key)))
+  (let* ((entry (bibtex-completion-get-entry key))
          (entry-type (bibtex-completion-get-value "=type=" entry)))
     (format "@%s{%s,\n%s}\n"
             entry-type key
@@ -854,12 +826,11 @@ defined.  Surrounding curly braces are stripped."
              concat
              (format "  %s = %s,\n" name value)))))
 
-(defun bibtex-completion-add-PDF-attachment (candidates)
+(defun bibtex-completion-add-PDF-attachment (keys)
   "Attach the PDFs of the selected entries where available."
   (--if-let
       (-flatten
-       (-map 'bibtex-completion-find-pdf
-             (bibtex-completion-normalize-candidates candidates)))
+       (-map 'bibtex-completion-find-pdf keys))
       (-each it 'mml-attach-file)
     (message "No PDF(s) found.")))
 
@@ -892,9 +863,9 @@ line."
         (delete-window window)
       (switch-to-buffer (other-buffer)))))
 
-(defun bibtex-completion-edit-notes (key)
-  "Open the notes associated with the entry using `find-file'."
-  (let ((key (bibtex-completion-normalize-candidate key)))
+(defun bibtex-completion-edit-notes (keys)
+  "Open the notes associated with the selected entries using `find-file'."
+  (dolist (key keys)
     (if (and bibtex-completion-notes-path
              (f-directory? bibtex-completion-notes-path))
                                         ; One notes file per publication:
@@ -936,11 +907,11 @@ line."
   (or (get-file-buffer file)
       (find-buffer-visiting file)))
 
-(defun bibtex-completion-show-entry (key)
-  "Show the entry in the BibTeX file."
+(defun bibtex-completion-show-entry (keys)
+  "Show the first selected entry in the BibTeX file."
   (catch 'break
     (dolist (bibtex-file (-flatten (list bibtex-completion-bibliography)))
-      (let ((key (bibtex-completion-normalize-candidate key))
+      (let ((key (car keys))
             (buf (bibtex-completion-buffer-visiting bibtex-file)))
         (find-file bibtex-file)
         (goto-char (point-min))