--- a/helm-bibtex.el
+++ b/helm-bibtex.el
@@ -170,6 +170,8 @@ it comes out in the right buffer."
 (helm-bibtex-helmify-action bibtex-completion-insert-key helm-bibtex-insert-key)
 (helm-bibtex-helmify-action bibtex-completion-insert-bibtex helm-bibtex-insert-bibtex)
 (helm-bibtex-helmify-action bibtex-completion-add-PDF-attachment helm-bibtex-add-PDF-attachment)
+(helm-bibtex-helmify-action bibtex-completion-edit-notes helm-bibtex-edit-notes)
+(helm-bibtex-helmify-action bibtex-completion-show-entry helm-bibtex-show-entry)
 
 ;; Helm sources:
 
@@ -186,8 +188,8 @@ it comes out in the right buffer."
              "Insert BibTeX key"          'helm-bibtex-insert-key
              "Insert BibTeX entry"        'helm-bibtex-insert-bibtex
              "Attach PDF to email"        'helm-bibtex-add-PDF-attachment
-             "Edit notes"                 'bibtex-completion-edit-notes
-             "Show entry"                 'bibtex-completion-show-entry))
+             "Edit notes"                 'helm-bibtex-edit-notes
+             "Show entry"                 'helm-bibtex-show-entry))
   "Source for searching in BibTeX files.")
 
 (defvar helm-source-fallback-options