Currently `helm-bibtex` caches unformatted entries and only formats the filtered ones through helm's `filtered-candidates-transformer` mechanism. `ivy-bibtex`, on the other hand, formats all entries before caching them.

This difference makes `ivy-bibtex` much slower than `helm-bibtex` for large bibliographies, and this PR fixes this by making `ivy-bibtex` cache unformatted entries and only format the displayed entries through ivy's `display-transformer` mechanism. There are some differences between helm's and ivy's mechanisms, which we have to take into account, but once we do that the two frontends consistently cache unformatted entries and pass list of keys to their actions, and they are similar in terms of speed. Here are the timings in seconds that I obtain with `crypto.bib` (same methodology as in #137 ):

|      | before | after |
| --- | --- | --- |
| `bibtex-completion-candidates` | 23.8 | 23.8 |
| `helm-bibtex` | 24.0 | 24.0 |
| `ivy-bibtex` | 38.6 | 24.0 |

There are five commits in the PR. The first two are the main ones that actually do the job. The last three are just cleanup that naturally follows from the first two, in my opinion, but I don't know if you will agree so I made separate commits. A detailed explanation follows. Of course let me know if you would like me to change some things.

The first commit makes `ivy-bibtex` cache unformatted entries (like `helm-bibtex`), and then only format the displayed entries through `ivy-set-display-transformer`. Unlike helm, ivy transforms only the displayed candidates and not all the filtered ones, and it transforms them one by one. Also, the transformer function only takes as argument the string to be formatted (that is, the `car` of the candidate), and outputs the formatted string to be displayed. So we have to recover the corresponding alist (the `cdr` of the candidate) from the unformatted string. ivy makes this fast and easy by storing the candidate's index in the candidates list as a text property in the unformatted string. So we simply get the index, fetch the corresponding entry, and use its alist to generate the formatted string. To do this efficiently I separated the formatting code out of the function `bibtex-completion-candidates-formatter` into a new function `bibtex-completion-format-entry`.

Another difference between ivy's and helm's mechanisms is that ivy only transforms the candidates for display but passes the untransformed selected candidate to actions. So `ivy-bibtex` now passes a cons cell (string . alist), instead of (string . key) previously (`helm-bibtex`, on the other hand, passes a list of keys). This makes `bibtex-completion-normalize-candidates` fail to correctly pass keys to actions. The second commit fixes this. But instead of tweaking `bibtex-completion-normalize-candidates` to figure out where to find the keys to pass without knowing which frontend is used, it seemed more natural to me to let each frontend do the job of extracting and passing keys to actions. `helm-bibtex-helmify-action` makes sure that `helm-bibtex` always passes a list of keys. I simply defined an `ivy-bibtex-ivify-actions` macro to make sure that `ivy-bibtex` also passes a similar list (with a single key, since ivy does not support selecting multiple candidates).

As I said, at this point everything works and the remaining commits just do some cleanup:
- The third one moves `bibtex-completion-candidates-formatter` to `helm-bibtex.el`, since only this frontend now uses it (the new function `bibtex-completion-format-candidate`,  which is used by both frontends, remains in `bibtex-completion.el`). In fact I merged this function into `helm-bibtex-candidates-formatter`.
- The fourth one removes the optional `formatter` argument to `bibtex-completion-candidates`, since it is no longer used.
- The fifth one removes `bibtex-completion-normalize-candidates`, which I think is no longer needed since both frontends now consistently pass list of keys to actions. It seems simpler to me to let the generic action functions in `bibtex-completion.el` assume that they systematically receive list of keys, and then let each frontend modify these functions as needed.

Sorry for the lengthy explanations, I hope at least they are clear.