@tmalsburg 2016-09-28
Thank you for the PR and the explanations.  It all looks very clean and well thought-out.  Just one comment:

In `ivy-bibtex-display-transformer`, you use `frame-width`.  If `frame-width` is like `window-width`, I think this won't work under some circumstances because it may give you more characters than actually fit on a line and this would cause unnecessary line breaks or continuation glyphs.  That's why I introduced `helm-bibtex-window-width` which at least takes the size of the fringes into account (but not the font size).  In Emacs 25, there is `window-max-chars-per-line` which is more precise (my first contribution to Emacs), but we can't rely on this function because it's too new. Perhaps move `helm-bibtex-window-width` to bibtex-completion and also use it in ivy-bibtex?