@ericdanan 2016-09-28
`frame-width` was used already in `ivy-bibtex-candidates-formatter`. I noticed that `helm-bibtex` was using the more sophisticated `helm-bibtex-window-width` and followed the link in the comment above this function, but I didn't fully understand the issues there. I supposed that the situation was probably simpler with `ivy-bibtex` (maybe because ivy always uses the minibuffer to display candidates) and this was probably why `frame-width` was simply used there.

Anyway, I don't know what the problematic circumstances may be, but for me `ivy-bibtex` perfectly fits candidates to the minibuffer width (and so does `helm-bibtex`, whether in full frame or not). But if I use `helm-bibtex-window-width` instead of `frame-width` in `ivy-bibtex`, then it no longer works when I have two windows side by side in the frame: the candidates in the minibuffer then only use the width of the window from which I called `ivy-bibtex` instead of the full minibuffer width.

So I don't know how to do better than `frame-width`...