@tmalsburg 2016-09-28
Ok, then lets stick with frame-width.