@ericdanan 2016-09-28
I would say it is ready to be merged. I have tested `ivy-bibtex` and `helm-bibtex` and tried the different actions, everything worked for me. The only problem I can think of is if users have customized the actions for `ivy-bibtex`, then it will stop working for them until they update their setup to use the "ivified" actions. I don't see how we can avoid this. Should we mention it in the `README`?