@tmalsburg 2016-09-29
Re ivyfied actions: Perhaps add a news item in `README.org` and one sentences saying that they have to use `ivy-bibtex-...` instead of `bibtex-completion-...` for actions.