@tmalsburg 2016-09-29
I changed the README and the added an item in the news section of ivy-bibtex.el.  Thank you for this PR, great work!